Split view file from dataObject file
